requirements.txt
# GameForge AI System Dependencies
# Production-ready package list with pinned versions for reproducible builds
# Last updated: September 13, 2025

# =====================================
# CORE PYTHON PACKAGES
# =====================================
pip==23.3.1
setuptools==68.2.2
wheel==0.41.3

# =====================================
# WEB FRAMEWORK & API
# =====================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
starlette==0.27.0
python-multipart==0.0.6
aiofiles==23.2.1

# =====================================
# HTTP CLIENT LIBRARIES
# =====================================
httpx==0.25.2
requests==2.31.0

# =====================================
# DATABASE & STORAGE
# =====================================
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
aiosqlite==0.21.0
alembic==1.13.0
redis==5.0.1

# =====================================
# BACKGROUND PROCESSING
# =====================================
celery==5.3.4
kombu==5.3.4

# =====================================
# AI/ML PIPELINE
# =====================================
torch==2.1.1
torchvision==0.16.1
diffusers==0.24.0
transformers==4.36.2
accelerate==0.25.0
xformers==0.0.22.post7
safetensors==0.4.1
huggingface-hub==0.19.4
datasets==2.14.6

# =====================================
# IMAGE PROCESSING
# =====================================
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.4
scipy==1.11.4
scikit-image==0.22.0

# =====================================
# AUTHENTICATION & SECURITY
# =====================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==41.0.8
PyJWT==2.8.0
bleach==6.1.0
authlib==1.2.1
httpx==0.25.2

# =====================================
# CLOUD STORAGE PROVIDERS
# =====================================
boto3==1.34.0
botocore==1.34.0
azure-storage-blob==12.19.0
google-cloud-storage==2.10.0
minio==7.2.0

# =====================================
# VAULT INTEGRATION
# =====================================
hvac==2.0.0

# =====================================
# MONITORING & OBSERVABILITY
# =====================================
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
structlog==23.2.0
python-json-logger==2.0.7
sentry-sdk==1.38.0

# =====================================
# OPENTELEMETRY
# =====================================
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-prometheus==1.12.0rc1
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-instrumentation-celery==0.42b0
opentelemetry-instrumentation-psycopg2==0.42b0
opentelemetry-instrumentation-logging==0.42b0
opentelemetry-propagator-b3==1.21.0
opentelemetry-propagator-jaeger==1.21.0

# =====================================
# CONFIGURATION & UTILITIES
# =====================================
python-dotenv==1.0.0

